Move geojson processing into product
